domus/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-ckeditor-5==0.2.18
django-filter==25.1
django-js-asset==3.1.2
djangorestframework==3.16.0
pillow==11.2.1
sqlparse==0.5.3
tzdata==2025.2
